requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Celery==5.2.1
redis==4.0.2
psycopg2-binary==2.9.3
pydantic==1.9.0
MarkupSafe==2.0.1
Flask-Testing
